+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.0.4] - Unreleased
+### Changed
+- Removed `includes` option for `bundler` plugin.
+  They are detected automatically now!
+
 ## [1.0.3] - 2021-08-30
 ### Added
 - Improved `bundler` plugin with the following additions:
@@ -1088,6 +1093,7 @@ The first version.
 [#125]: https://github.com/lumeland/lume/issues/125
 [#131]: https://github.com/lumeland/lume/issues/131
 
+[1.0.4]: https://github.com/lumeland/lume/compare/v1.0.3...HEAD
 [1.0.3]: https://github.com/lumeland/lume/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/lumeland/lume/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/lumeland/lume/compare/v1.0.0...v1.0.1